Add code coverage hook
