+++ scripts/after-code-coverage-success.js
@@ -1,0 +1,7 @@
+async function runAsync() {
+  require('./test-schematics');
+}
+
+module.exports = {
+  runAsync
+};
