+++ scripts/build-schematics.js
@@ -1,15 +1,7 @@
-const crossSpawn = require('cross-spawn');
 const fs = require('fs-extra');
 const path = require('path');
 
-const LIB_PATH = path.join(process.cwd(), 'projects/i18n');
-
-function runCommand(command, args) {
-  crossSpawn.sync(command, args, {
-    cwd: LIB_PATH,
-    stdio: 'inherit'
-  });
-}
+const runCommand = require('./run-command');
 
 function buildSchematics() {
   console.log('Building library schematics...');
